% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_SpatialDLS_v2.R
\name{dimreduc.spot}
\alias{dimreduc.spot}
\title{Dimension Reduction of the transcriptomic original data}
\usage{
dimreduc.spot(so, method_reduction, n.dim, max_gamma, genes.use)
}
\arguments{
\item{so}{a Seurat object}

\item{method_reduction}{a string indicating the method. Choose between "log_normalize", "no" and "SCT".}

\item{n.dim}{number of principal components to use from PCA}

\item{max_gamma}{value of the maximum reachable gamma}

\item{genes.use}{genes to use when "method_reduction = 'genes.list'"}
}
\value{
a Seurat object
}
\description{
This function performs dimensionality reduction of the transcriptomic original data based on the method asked by the user. The available methods are "PCA", "NMF" and "genes.list" for using directly the expression of selected genes instead of other methods for dimensionality reduction.
}
