% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_SpatialDLS_v2.R
\name{neighbor_graph_v2}
\alias{neighbor_graph_v2}
\title{Creation of the KNN graph to links the spots based on the spatial proximity}
\usage{
neighbor_graph_v2(
  spotPositions,
  col.x,
  col.y,
  k.spots,
  countMatrix,
  n.dim,
  method_similarity,
  method_knn,
  k.knn,
  sigs,
  method_normalization,
  split_not_connected,
  split_annotation,
  split_vector,
  n.cpu,
  plot.graph,
  pct,
  dist.thresh,
  method_reduction,
  genes.use
)
}
\arguments{
\item{spotPositions}{a data.frame of the coordinates of the spots}

\item{col.x}{a character of the name of the x coordinates to get from the coordinates data frame}

\item{col.y}{a character of the name of the y coordinates to get from the coordinates data frame}

\item{k.spots}{number of neighborhoods to take into account when choosing the second method to build the KNN}

\item{countMatrix}{raw matrix of the gene expression for each spot}

\item{n.dim}{number of principal components to use from PCA}

\item{method_similarity}{method to compute transcriptional similarity from distance}

\item{method_knn}{method to compute the KNN
\itemize{
\item "1" - KNN is built with RANN::nn2() function (default)
\item "2" - KNN is built by computing every distances between each spot and filtered by choosing the number of neighborhoods to take into account
}}

\item{k.knn}{numbers of neighboring spots to compute for each spot}

\item{sigs}{parameter when computing the transcriptional similarity with method "2"}

\item{method_normalization}{method for normalization of the raw count matrix
\itemize{
\item "log_normalize" for a Seurat Log Normalization
\item "SCT" for a Seurat SCT Normalization
\item "no" for using raw counts
}}

\item{split_not_connected}{split or not the spots in the KNN graph if their distance is too far compared to the others}

\item{split_annotation}{split or not the spots in the KNN graph based on their annotation}

\item{split_vector}{vector of the spots annotations on which the KNN will be split}

\item{n.cpu}{number of cpu to use during parallelized computation of distances. By default, the maximum amount of cpu available is chosen automatically. But if your computer doesn't support it, please specify your desired number of cpu.}

\item{plot.graph}{boolean of whether or not plotting the KNN graph}

\item{pct}{percentage of the connections to keep
\itemize{
\item "1" - similarity = 1 - (distance/max(distance))
\item "2" - similarity = exp(-(distance^2)/sigma)
\item "3" - similarity = proxy::pr_dist2simil(distance)
}}

\item{dist.thresh}{fixed value to use instead of using a percentage to filter out connections}

\item{method_reduction}{a string indicating the method. Choose between "log_normalize", "no" and "SCT".}

\item{genes.use}{genes to use when "method_reduction = 'genes.list'"}
}
\value{
a list with components
\itemize{
\item graph - an igraph graph
\item so - a seurat object
}
}
\description{
This function builds the KNN graph to links the spots based on the spatial proximity
}
