% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_SpatialDLS_v2.R
\name{normalize.spot}
\alias{normalize.spot}
\title{Normalization of the transcriptomic original data}
\usage{
normalize.spot(so, method_normalization)
}
\arguments{
\item{so}{a Seurat object}

\item{method_normalization}{a string indicating the method. Choose between "log_normalize", "no" and "SCT".}
}
\value{
a Seurat object
}
\description{
This function normalize the transcriptomic original data based on the method asked by the user. The available methods are "log_normalize", "no" for using the raw counts and "SCT".
}
