% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_SpatialDLS.R
\name{shape_metrics}
\alias{shape_metrics}
\title{Computation of the geometric values of metaspots}
\usage{
shape_metrics(MC, polygon_col, membership_col)
}
\arguments{
\item{MC}{Metaspot object obtained from SCimplify_SpatialDLS function}

\item{polygon_col}{character of the name where the polygons representing the metaspots are stored in the MC object}

\item{membership_col}{character of the name where the memberships of the spots are stored in the MC object}
}
\value{
a dataframe with columns
\itemize{
\item memberships - memberships of the metaspots
\item elongations - elongations of the metaspots
\item circularities - circularities of the metaspots
\item convexities - convexities of the metaspots
}
}
\description{
This function computes the geometric values of metaspots based on the previously computed polygons using pliman R package
}
