% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSTplot.R
\name{SpatialDimPlotSC}
\alias{SpatialDimPlotSC}
\title{Plotting metaspots on the original spots}
\usage{
SpatialDimPlotSC(
  original_coord,
  MC,
  sc.col = NULL,
  sc.col2 = NULL,
  col.x = "imagecol",
  col.y = "imagerow",
  polygons_col,
  alpha = 1,
  meta_data,
  alpha_hull = 0.5,
  spot.color = NULL
)
}
\arguments{
\item{original_coord}{dataframe of the coordinates of the original spots}

\item{MC}{Metaspot object obtained from SCimplify_SpatialDLS function}

\item{sc.col}{a character of the name of the annotation of the spots to get in the meta.data}

\item{sc.col2}{a character of the name of the corresponding membership of the spots to get in the meta.data}

\item{col.x}{a character of the name of the x coordinates to get from the coordinates data frame}

\item{col.y}{a character of the name of the y coordinates to get from the coordinates data frame}

\item{polygons_col}{a character of the name where the polygons have been stored in the MC object}

\item{alpha}{parameter of the alpha for the spots}

\item{meta_data}{dataframe containing both the annotations of the spots and their membership}

\item{alpha_hull}{parameter of the alpha for the polygons of the metaspots}

\item{spot.color}{vector for the spot colors}
}
\value{
a ggplot2 plot
}
\description{
This function projects the metaspots on the original spots based on the spatial coordinates
}
