% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_SpatialDLS.R
\name{split_unconnected}
\alias{split_unconnected}
\title{Split metaspots that are unconnect in KNN graph}
\usage{
split_unconnected(MC)
}
\arguments{
\item{MC}{Metaspot object obtained from SCimplify_SpatialDLS function}
}
\value{
a vector of membership with updated split memberships
}
\description{
This function splits metaspots that are unconnect in KNN graph
}
