% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_SpatialDLS.R
\name{supercell_2_Seuratv5}
\alias{supercell_2_Seuratv5}
\title{Superspots to Seurat object}
\usage{
supercell_2_Seuratv5(SC.GE, SC, fields = c())
}
\arguments{
\item{SC.GE}{gene expression matrix with genes as rows and cells as columns}

\item{SC}{super-cell (output of \link{SCimplify} function)}

\item{fields}{which fields of \code{SC} to use as cell metadata}
}
\value{
\link[Seurat]{Seurat} object
}
\description{
This function transforms superspot gene expression and superspot partition into \link[Seurat]{Seurat} object
}
