% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSTplot.R
\name{supercell_metaspots_shape_v2}
\alias{supercell_metaspots_shape_v2}
\title{Creation of polygons around the metaspots (upgrade)}
\usage{
supercell_metaspots_shape_v2(
  MC,
  spotpositions,
  annotation,
  concavity,
  membership_name = "membership"
)
}
\arguments{
\item{MC}{Metaspot object obtained from SCimplify_SpatialDLS function}

\item{spotpositions}{a data.frame of the coordinates of the spots}

\item{annotation}{a vector of the annotation for each spot}

\item{concavity}{a number to define the concavity of the polygons}

\item{membership_name}{a character of the name where the memberships have been stored in the MC object}
}
\value{
a dataframe with columns
\itemize{
\item x - x coordinates of the polygons
\item y - y coordinates of the polygons
\item cell_type - annotation of the polygons
\item membership - membership of the polygons
}
}
\description{
This function detects creates of around the metaspots based of the memberships of spots
}
