% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify_SpatialDLS.R
\name{supercell_spatial_centroids}
\alias{supercell_spatial_centroids}
\title{Computation of the spatial centroids about metaspots}
\usage{
supercell_spatial_centroids(MC, spotPositions)
}
\arguments{
\item{MC}{Metaspot object obtained from SCimplify_SpatialDLS function}

\item{spotPositions}{a data.frame of the coordinates of the spots}
}
\value{
a dataframe of the centroids coordinates for each metaspot
}
\description{
This function computes the spatial centroids of metaspots from the coordinates of the original spots
}
